chore: organise handler test files
