+++ packages/adapters/test/handlers/adapters/metadata.test.ts
@@ -2,8 +2,8 @@ import { HttpResponse, mockFetch } from "#msw-utils";
 import { createCache } from "@mojis/internal-utils";
 import { emojiTest } from "@mojis/loomicode";
 import { afterEach, describe, expect, it } from "vitest";
-import { handler } from "../../src/handlers/adapter/metadata";
-import { cleanupAdapterTest, setupAdapterTest } from "../test-utils";
+import { handler } from "../../../src/handlers/adapter/metadata";
+import { cleanupAdapterTest, setupAdapterTest } from "../../test-utils";
 
 describe("metadata adapter handler", () => {
   const mockContext = {
