+++ packages/adapters/test/handlers/adapters/unicode-names.test.ts
@@ -1,8 +1,8 @@
 import { HttpResponse, mockFetch } from "#msw-utils";
 import { createCache } from "@mojis/internal-utils";
 import { afterEach, describe, expect, it } from "vitest";
-import { handler } from "../../src/handlers/adapter/unicode-names";
-import { cleanupAdapterTest, setupAdapterTest } from "../test-utils";
+import { handler } from "../../../src/handlers/adapter/unicode-names";
+import { cleanupAdapterTest, setupAdapterTest } from "../../test-utils";
 
 describe("unicode-names adapter handler", () => {
   const mockContext = {
