+++ packages/adapters/test/handlers/adapters/variations.test.ts
@@ -2,8 +2,8 @@ import { HttpResponse, mockFetch } from "#msw-utils";
 import { createCache } from "@mojis/internal-utils";
 import { variations } from "@mojis/loomicode";
 import { afterEach, describe, expect, it } from "vitest";
-import { handler } from "../../src/handlers/adapter/variations";
-import { cleanupAdapterTest, setupAdapterTest } from "../test-utils";
+import { handler } from "../../../src/handlers/adapter/variations";
+import { cleanupAdapterTest, setupAdapterTest } from "../../test-utils";
 
 describe("variations adapter handler", () => {
   const mockContext = {
