+++ packages/adapters/test/handlers/composite/emoji-data.test.ts
@@ -1,0 +1,5 @@
+import { expect, it } from "vitest";
+
+it("should handle emoji data", () => {
+  expect(true).toBe(true);
+});
